# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "ZooClick",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/{user}/create-channel": {
            "post": {
                "tags": [
                    "Channel"
                ],
                "summary": "Создание канала",
                "description": "Создание канала",
                "operationId": "CreateChannel",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Название канала",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Описание канала",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "Изображение для канала",
                        "required": false,
                        "schema": {
                            "type": "file"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/{user}/channels": {
            "get": {
                "tags": [
                    "Channel"
                ],
                "summary": "Получение каналов",
                "description": "Получение каналов",
                "operationId": "GetChannel",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/{user}/{model}/{modelId}/like": {
            "post": {
                "tags": [
                    "Like"
                ],
                "summary": "Лайк сущности",
                "description": "Лайк сущности",
                "operationId": "LikeTheme",
                "parameters": [
                    {
                        "name": "user",
                        "in": "query",
                        "description": "id пользователя",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "model",
                        "in": "query",
                        "description": "Название модели",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "modelId",
                        "in": "query",
                        "description": "id записи которую лайкаем",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/{channel}/create-theme": {
            "post": {
                "tags": [
                    "Theme"
                ],
                "summary": "Создание темы",
                "description": "Создание новой темы",
                "operationId": "CreateTheme",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Создание темы",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Описание темы",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "images[]",
                        "in": "query",
                        "description": "Дополнителыние изображения",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "comments_active",
                        "in": "query",
                        "description": "Статус комментариев",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "repost_id",
                        "in": "query",
                        "description": "Тема с репостом",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Регистрация нового юзера",
                "description": "Регистрация",
                "operationId": "UserRegister",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Имя юзера",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email юзера",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password юзера",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login юзера",
                "description": "Логин",
                "operationId": "UserLogin",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email юзера",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password юзера",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}